rename maps to config
