+++ src/games/RPG/config.tsx
